refactor FnvMap to it's own module
