+++ vm/src/types.rs
@@ -1,6 +1,7 @@
 use base::symbol::{Symbol, SymbolRef};
 use base::types;
-use base::types::{Alias, KindEnv, TypeEnv, TcType, Type, FnvMap};
+use base::types::{Alias, KindEnv, TypeEnv, TcType, Type};
+use base::fnv::FnvMap;
 
 pub use self::Instruction::*;
